requirement.txt
# 现有依赖
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
python-dotenv>=1.0.0
pyyaml>=6.0
redis>=5.0.0
pymysql>=1.1.0

langchain>=0.1.0
langchain-openai>=0.0.5  
langchain-core>=0.1.0